+++ Backend/src/main/java/main/services/PaymentTransactionService.java
@@ -3,13 +3,14 @@ package main.services;
 import main.dtos.response.PaymentTransactionResponse;
 import main.enums.PaymentPurpose;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public interface PaymentTransactionService {
 
     // Create a PayOS payment for a booking
-    PaymentTransactionResponse createPayment(UUID bookingId, PaymentPurpose paymentPurpose, Long amount);
+    PaymentTransactionResponse createPayment(UUID bookingId, PaymentPurpose paymentPurpose, BigDecimal amount);
 
     // Get payment transaction by id
     PaymentTransactionResponse getById(UUID paymentTransactionId);
